server/requirements.txt
asgiref==3.6.0
black==23.1.0
click==8.1.3
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
flake8==6.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.1.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
